+++ src/app/(Sub)/bookmark/page.tsx
@@ -1,19 +1,11 @@
 import EmptyCase from '@/components/Feed/EmptyCase';
 import PhotoList from '@/components/Feed/PhotoList';
-import { Header } from '@/components/Header';
-import CloseButton from '@/components/Header/CloseButton';
+import SubHeader from '@/components/Header/SubHeader';
 
 export default function BookmarkPage() {
   return (
     <>
-      <Header
-        leftNode={
-          <div className="g-12 itmes-center flex">
-            <CloseButton />
-            <h4>북마크</h4>
-          </div>
-        }
-      />
+      <SubHeader title="북마크" />
       <EmptyCase
         title={'포즈를 보관해 보세요!'}
         text={`북마크 버튼으로 포즈를 보관할 수 있어요.\n포즈피드에서 멋진 포즈를 찾아 보관해 보세요.`}
