+++ src/app/(Sub)/menu/page.tsx
@@ -1,12 +1,12 @@
 import LoginSection from './components/LoginSection';
 import MakerSection from './components/MakerSection';
-import MenuHeader from './components/MenuHeader';
 import MenuListSection from './components/MenuListSection';
+import SubHeader from '../../../components/Header/SubHeader';
 
 export default function MenuPage() {
   return (
     <div className="px-20">
-      <MenuHeader />
+      <SubHeader title="메뉴" />
       <LoginSection />
       <MenuListSection />
       <MakerSection />
